WIP: Social Interactions refactor
