+++ src/routes/collection.$collectionId.tsx
@@ -4,7 +4,7 @@ import { Nip05Badge } from '@/components/Nip05Badge.tsx'
 import { ProductCard } from '@/components/ProductCard'
 import { Button } from '@/components/ui/button'
 import { UserCard } from '@/components/UserCard'
-import { ZapButton } from '@/components/ZapButton.tsx'
+import { ZapButton } from '@/components/social/ZapButton'
 import { useBreakpoint } from '@/hooks/useBreakpoint'
 import { useEntityPermissions } from '@/hooks/useEntityPermissions'
 import { ndkActions } from '@/lib/stores/ndk'
